requirements.txt
boto3==1.28.20
botocore==1.31.20
jmespath==1.0.1
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
